Refactored SanPanel
